Remove Cage Porosity models
